refactor: Extract into new service `Exporters::Antweb::ExportTaxon`
