+++ spec/lib/formatters/advanced_search_html_formatter_spec.rb
@@ -2,6 +2,8 @@ require 'spec_helper'
 
 class FormattersAdvancedSearchHtmlFormatterTestClass
   include Formatters::AdvancedSearchHtmlFormatter
+  include ActionView::Helpers::TagHelper # For `#content_tag`.
+  include ActionView::Context # For `#content_tag`.
 end
 
 describe Formatters::AdvancedSearchHtmlFormatter do
